Flip Learning

+++ 02/assignment/waffleblog/posts/views.py
@@ -1,9 +1,13 @@
 from django.shortcuts import render, redirect
-from .models import Post
-from .forms import PostForm, DeleteForm
+from .models import Post, Comment
+from .forms import PostForm, SignupForm, CommentForm
+from braces.views import LoginRequiredMixin
 
 # Create your views here.
 
+def index(request):
+    print(request.user.is_authenticated)
+    return render(request, "posts/index.html")
 
 def posts_list(request):
     posts = Post.objects.all()
@@ -13,15 +17,28 @@ def posts_list(request):
 
 def post_detail(request, post_id):
     post = Post.objects.get(id=post_id)
-    context = {"post": post}
-    return render(request, "posts/post_detail.html", context)
+    comment = Comment.objects.filter(post=post)
+    if request.method == "POST":
+        comment_form = CommentForm(request.POST)
+        if comment_form.is_valid():
+            new_comment = comment_form.save(commit=False)
+            new_comment.post = post
+            new_comment.save()
+        return redirect("post_detail", post_id=post.id)
+    else:
+        comment_form = CommentForm()
+        context = {"post": post, "comment": comment, "form": comment_form}
+        return render(request, "posts/post_detail.html", context)
 
 
 def post_create(request):
+    if not request.user.is_authenticated:
+        return redirect("posts_list")
     if request.method == "POST":
         post_form = PostForm(request.POST)
         if post_form.is_valid():
             post = post_form.save()
+            post.save()
         return redirect("post_detail", post_id=post.id)
     else:
         post_form = PostForm()
@@ -30,6 +47,8 @@ def post_create(request):
 
 def post_update(request, post_id):
     post = Post.objects.get(id=post_id)
+    if not request.user.is_authenticated:
+        return redirect("post_detail", post_id=post.id)
     if request.method == "POST":
         post_form = PostForm(request.POST, instance=post)
         if post_form.is_valid():
@@ -42,9 +61,21 @@ def post_update(request, post_id):
 
 def post_delete(request, post_id):
     post = Post.objects.get(id=post_id)
+    if not request.user.is_authenticated:
+        return redirect("post_detail", post_id=post.id)
     if request.method == 'POST':
         post.delete()
         return redirect("posts_list")
     else: 
         return render(request, 'posts/post_confirm_delete.html', {'post': post})
     
+
+def change_profile(request):
+    if request.method == "POST":
+        form = SignupForm(request.POST)
+        if form.is_valid():
+            form.signup(request, form)
+            return redirect("index")
+    else:
+        form = SignupForm()
+        return render(request, 'posts/change_profile.html', {'form': form})
